Refactoring of methods in classes and some tweaks.
Create new classes to separate some responsibilities
