+++ mathGame.batista92/mathGame.batista92/Models/Game.cs
@@ -1,0 +1,10 @@
+﻿namespace mathGame.batista92.Models;
+
+internal class Game
+{
+    public DateTime Date { get; set; }
+
+    public int Score { get; set; }
+
+    public string Type { get; set; }
+}
